requirements.txt
mistralai
requests
beautifulsoup4
pydantic
python-dotenv
